metrics_agent/requirements.txt
psutil==5.9.8
redis==5.0.4
python-dotenv==1.0.1
prometheus-client==0.20.0
